refactor: Rename redis handler to redis pool.
